Move all encode[proofName]Transaction functions from the protocol/.../ACE/validator test files to aztec.js/.../proof/[proofName]
Done to give better compartmentalisation of proof related functions

Remove mint tests taken from joinSplit - a future commit will implement mint compatible tests
